Adjusted curves in humidity and temperature simulations so that groups correspond closer to Earth norms.
Added some weighting variables in humidity simulation to make it easier to adjust output. Increased the weighting of latitude for determining temperature in temperature simulation to produce a result that seems closer to Earth's (less chaotic).

Added an error catch to astar.py to prevent a CTD error that seemed to occur on maps of 1024x512 and larger.

Added a variable called latshift to generation.py. This will ultimately allow people to rerun a specific seed but shift the continental position up and down. This way if you identify a spot you think would be a good central location but you think that the campaign/story would benefit from having it further north or south the simulation can be rerun and the results shifted.

Reversed the order of generating temperature and humidity in generation.py as the temperature has an effect on the humidity.

Added biomes.scm, a script for GIMP that should create layers and layer groups for the biome png files.