Modification to the routines for drawing colored elevation maps.
This raises the minimum altitude of the land to sea level and then normalizes the elevation values to the range of 1 to 12. This prevents the need of repeating color bands to represent higher ranges. As elevation has no inherent value the normalization causes no real distortions. The raising of the minimum land value means there is no land with an altitude above sea level (something which can occur in the real world) but it makes the elevation map easier to import into other programs such as Fractal Terrains and such a solution is already used in the grayscale height maps.