Added a bit more code to draw_scatter_plot. It's ugly, boneheaded, cut and paste but I'm just a little too tired right now to create a more elegant looping system.
The code will color cells that 'should not occur' neutral gray. This makes it easier to see when results are spilling over into those areas. It's fine to have some spill as the biome will simply be set to the lowest 'good' value but the graph now definitely lets you see when there are large gaps or overlays.