Update __init__.py - resolve Windows error 32
When running export in Windows, errors out: "WindowsError: [Error 32] The process cannot access the file because it is in use by another process: path_to_temp_file"

Flushing (new line 132) and adding a call to close with the file handle (new line 134) resolves the error. Just flushing, or just closing, does not.

Thanks!